ref(js): Factor button functionality
- Move `useButtonFunctionality` into it's own module, keeping index.tsx
  specifc to JUST what is required to render the Button, and not sharing
  anything with linkButton

- Move common styles out of the index module into styles.tsx, renames
  index.chonk to styles.chonk as well.

- Move common types into types.tsx

- Renames `getChonkButtonStyles` to `DO_NOT_USE_getChonkButtonStyles` to
  match the `DO_NOT_USE_getButtonStyles`
