+++ static/app/components/core/button/styles.chonk.tsx
@@ -1,6 +1,6 @@
 import type {DO_NOT_USE_ChonkTheme} from '@emotion/react';
 
-import type {ButtonProps} from 'sentry/components/core/button';
+import type {DO_NOT_USE_ButtonProps as ButtonProps} from 'sentry/components/core/button/types';
 import type {StrictCSSObject} from 'sentry/utils/theme';
 
 // @TODO: remove Link type in the future
@@ -33,7 +33,7 @@ function chonkPriorityToType(priority: ButtonProps['priority']): ChonkButtonType
   }
 }
 
-export function getChonkButtonStyles(
+export function DO_NOT_USE_getChonkButtonStyles(
   p: Pick<ButtonProps, 'size' | 'priority' | 'busy' | 'disabled' | 'borderless'> & {
     theme: DO_NOT_USE_ChonkTheme;
   }
